* master: (315 commits)
  svn update
  update addon_contrib scripts
  update addon scripts
  update translation files
  workarond for outliner group object selection trick - which is so slow it had to be disabled. Now use the first object of the group, which isn't great but is at least usable.
  correction to own commit - thanks Anthony Edlin for pointing out the error.
  Cycles XML: * Add recent integrator options.
  fix for segfault loading a file with a proxy that is a missing link.
  Refactoring of tiles opencl implementation: - Moved methods from NodeOperation to OpenCLDevice - Added check on Nvidia for local size
  avoid calls to BLF_height_default in a loop calling with the same value, also quiet warning.
  make mask handles draw with an outline - when outline option is enabled.
  fix for use of 2 uninitialized vars in the tiles compositor.
  fix for error with sequencer stip text drawing - was using unclamped handle sizes for the text bounds, this could be very wrong with long strips.
  * reset OpenCL innerloop size to 32.
  style cleanup
  move the frame length to the end for sequence drawing (mango request, since often the name is clipped out entirely so only then length is seen)
  Fix crash in compositing nodes with a node group with missing datablock, can happen with library linking and missing files.
  Fix #31883: cycles vectors pass rendering crash with some render layers using it and others not.
  Fix #31852: Sequencer duplicate cant move strips on Y axis in a meta
  removing from group now shows menu to select group to remove.
  ...

+++ intern/cycles/kernel/svm/svm_gradient.h
@@ -24,9 +24,9 @@ __device float svm_gradient(float3 p, NodeGradientType type)
 {
 	float x, y, z;
 
-	x= p.x;
-	y= p.y;
-	z= p.z;
+	x = p.x;
+	y = p.y;
+	z = p.z;
 
 	if(type == NODE_BLEND_LINEAR) {
 		return x;
@@ -45,7 +45,7 @@ __device float svm_gradient(float3 p, NodeGradientType type)
 		return (x + y)/2.0f;
 	}
 	else if(type == NODE_BLEND_RADIAL) {
-		return atan2(y, x)/(2.0f*M_PI_F) + 0.5f;
+		return atan2f(y, x) / (2.0f * M_PI_F) + 0.5f;
 	}
 	else {
 		float r = fmaxf(1.0f - sqrtf(x*x + y*y + z*z), 0.0f);
